Fix tim depedencies
f-strings were added in Python 3.6. Uses of this feature are replaced
with the more traditional .format() to allow tim to run on Xenial
without requiring the use of a backported Python.

Pyyaml was missing as a declared dependency.

As the README.md file is utf-8, avoid package installation problems by
explicitly setting the open() encoding type in the custom read()
function in setup.py. This helps for installation in docker containers
where the locale.getpreferredencoding() might return something exotic.